requirements.txt
lightgbm>=4.0
psycopg2-binary>=2.9
sqlalchemy>=2.0
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
tqdm>=4.65
optuna>=3.0
pyarrow>=12.0
